fix(web): guardrails server-only client import — next build failure
AgentGuardrailsCard.tsx ('use client', from #1710) value-imported
AGENT_GUARDRAIL_ACTION_TYPES from the server-only lib/api/agents.ts,
pulling the server-only guard into the client bundle → next build
failed with "importing a module that depends on server-only". Extracted
the pure guardrail types + action-type tuple into a client-safe
lib/api/agents.shared.ts; agents.ts re-exports them (server callers
unchanged); the card imports from the shared module. Same fix pattern
as goals.shared.ts (#1709). Verified: next build compiles clean.
