+++ apps/web/src/lib/api/agents.ts
@@ -20,34 +20,18 @@ export type AgentIdleBehavior = 'propose' | 'sleep' | 'self-improve';
 export type AgentFileName = 'SOUL.md' | 'AGENTS.md' | 'HEARTBEAT.md' | 'TOOLS.md' | 'agent.yml';
 
 // ── Agent Dispatch Guardrails ──
-// Mirrors `AgentGuardrails` (packages/agent/src/agents/guardrails.ts)
-// and the proposal action types
-// (packages/agent/src/entities/agent-action-proposal.entity.ts).
-
-export type AgentGuardrailActionType =
-    | 'spawn_agent'
-    | 'schedule_task'
-    | 'send_message'
-    | 'budget_override'
-    | 'other';
-
-export const AGENT_GUARDRAIL_ACTION_TYPES: readonly AgentGuardrailActionType[] = [
-    'spawn_agent',
-    'schedule_task',
-    'send_message',
-    'budget_override',
-    'other',
-] as const;
-
-export type AgentGuardrailsMode = 'require_approval' | 'autonomous';
-
-export interface AgentGuardrails {
-    mode: AgentGuardrailsMode;
-    /** Autonomous-mode narrowing; omitted = every unflagged type may auto-approve. */
-    autoApproveActionTypes?: AgentGuardrailActionType[];
-    /** Action types the Agent may never take (auto-rejected with an audit row). */
-    blockedActionTypes?: AgentGuardrailActionType[];
-}
+// Pure guardrail types + the action-type tuple live in a
+// `server-only`-free module (`agents.shared.ts`) so `'use client'`
+// components (e.g. AgentGuardrailsCard) can import them without pulling
+// this server-only module into the client bundle. Re-exported here so
+// server-side callers keep one import site.
+export {
+    AGENT_GUARDRAIL_ACTION_TYPES,
+    type AgentGuardrailActionType,
+    type AgentGuardrailsMode,
+    type AgentGuardrails,
+} from './agents.shared';
+import type { AgentGuardrails } from './agents.shared';
 
 export interface AgentPermissions {
     canCreateAgents: boolean;
